Moved factories into subfolders
