+++ cypress/factories/dpl-cms/getDplCmsPrivateConfiguration.ts
